refactor filenames for consistency
