+++ src/enums/operationType.ts
