+++ src/enums/roomType.ts
@@ -1,0 +1,3 @@
+export enum RoomType {
+  Normal = 1
+}
