--- src/enums/roomtype.ts
@@ -1,3 +1,0 @@
-export enum Roomtype {
-  Normal = 1
-}
