+++ src/managers/_manager.ts
